backend/requirements.txt
Flask==1.1.2
pandas
requests
python-dateutil==2.8.1
json5==0.9.5
jsonschema==3.2.0
